Rename cards_learning to cards_interval_learning
